requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
googlesearch-python==2020.0.2
html-text==0.5.2
idna==2.10
insta-scrape==0.2.1
instasave==0.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
loguru==0.4.1
lxml==4.5.2
MarkupSafe==1.1.1
multidict==4.7.6
pyinsta3==1.2
requests==2.24.0
soupsieve==2.0.1
tqdm==4.49.0
urllib3==1.25.9
Werkzeug==1.0.1
xmltodict==0.12.0
yarl==1.5.1
Pillow==7.2.0
pytesseract==0.3.6
Cython==0.29.21